+++ Frends.LDAP.CreateUser/Frends.LDAP.CreateUser/Definitions/Input.cs
@@ -51,7 +51,8 @@ public class Input
     public bool SetPassword { get; set; }
 
     /// <summary>
-    /// Password can be added to the unicodePwd field as a byte[] type.   
+    /// Password can be added to the unicodePwd field as a byte[] type
+    /// Requires connection.SecureSocketLayer or connection.TLS.
     /// </summary>
     /// <example>true</example>
     [DefaultValue(false)]