+++ Frends.LDAP.CreateUser/CHANGELOG.md
@@ -2,8 +2,7 @@
 
 ## [1.1.0] - 2025-09-26
 ### Updated
-- Input button added.
-- Task unicodes the passwords.
+- Added option to set the password for the new user in Unicode.
 
 ## [1.0.0] - 2022-09-23
 ### Added